Test - Patient

+++ app/src/main/java/com/example/crud_pacientes_e_medicos/model/Doctor.kt
@@ -6,7 +6,7 @@ import androidx.room.*
 data class Doctor(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "doctor_id")
-    var idDoctor: Long = 0,
+    var idDoctor: Int = 0,
 
     @ColumnInfo(name = "doctor_name")
     var nameDoctor: String,
